feat(render): #3 modularize scripts
