+++ src/index.mts
@@ -1,0 +1,15 @@
+/**!
+ * 
+ * render.js
+ * 
+ * Copyright (C) 2025 よね/Yone
+ * 
+ * Licensed under the MIT License.
+ * 
+ * https://github.com/yone1130/render.js
+ * 
+ */
+
+import { Render } from "./render.mjs";
+
+export default Render;
