+++ src/types/render-arguments.mts
@@ -1,0 +1,16 @@
+/**!
+ * 
+ * render.js
+ * 
+ * Copyright (C) 2025 よね/Yone
+ * 
+ * Licensed under the MIT License.
+ * 
+ * https://github.com/yone1130/render.js
+ * 
+ */
+
+export interface RenderArguments {
+    target: HTMLElement;
+    children: Array<HTMLElement>;
+}
